+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `ui.emojis` config option to turn off all emojis in the UI.
+- Clipboard support is now optional and off by default on Android/Termux, where it can't build anyway.
 
 ### Dependencies